Basic, lazy semver trick to avoid double linking
This change implements a lazy version of semver trick, reusing symbols
from secp256k1-sys crate. This avoid attempts to link libsecp256k1 twice
if two versions (0.15 and 0.17) of this crate are used. This was the
case in LNP/BP node implementation, for instance.

This particular instance of semver trick is considered "lazy" because it
only deduplicates the code related to the C library. It makes NO effort
to deduplicate the Rust code. This is already useful, because the crates
at least compile, even though they may get bloated. A more refined
version of this trick could be implemented in the future (as 0.16.7)
that would deduplicate the Rust code too.

This commit intentionally leaves the old code commented to make it
easier to compare the reexported items to the old ones. It will be
removed in the following commit.